pyproject.toml
[tool.poetry]
name = "classification"
version = "0.1.0"
description = ""
authors = ["Artem Verbov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
omegaconf = ">=2.3.0"
pydantic = ">=2.1.1"
torchmetrics = ">=1.1.1"
lightning = ">=2.0.6"
opencv-python-headless = "^4.7.0.72"
torchvision = ">=0.15.0"
albumentations = ">=1.3.1"
timm = "^0.9.7"
pytorch-lightning = "^2.0.8"
tensorboard = "^2.14.0"
torchinfo = "^1.8.0"
matplotlib = "^3.8.0"
pandas = "^2.1.0"
seaborn = "^0.12.2"
clearml = "^1.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

